+++ crates/config/src/lib.rs
@@ -17,12 +17,13 @@ pub mod internal;
 /// Built-in `merge` functions.
 pub mod merge;
 
-/// Utilities for generating schemas.
-pub mod schema;
+/// Renderers for rendering schema output.
+#[cfg(feature = "schema")]
+pub mod renderers;
 
-/// Utilities for generating TypeScript types.
-#[cfg(feature = "typescript")]
-pub mod typescript;
+/// Generate schemas for config and Rust types.
+#[cfg(feature = "schema")]
+pub mod schema;
 
 /// Built-in `validate` functions.
 pub mod validate;