+++ crates/config/src/renderers/mod.rs
@@ -1,0 +1,3 @@
+/// Renders TypeScript types.
+#[cfg(feature = "typescript")]
+pub mod typescript;