+++ crates/config/src/schema/mod.rs
@@ -1,1 +1,6 @@
+mod generator;
+mod renderer;
+
+pub use generator::*;
+pub use renderer::*;
 pub use schematic_types::*;