+++ crates/config/tests/defaults_test.rs
@@ -142,13 +142,14 @@ fn generates_typescript() {
     let sandbox = create_empty_sandbox();
     let file = sandbox.path().join("config.ts");
 
-    let mut generator = typescript::TypeScriptGenerator::new(file.clone());
-    generator.add::<NativeDefaults>();
+    let mut generator = schema::SchemaGenerator::default();
     generator.add::<CustomDefaults>();
     generator.add::<ReqOptDefaults>();
     generator.add::<ContextDefaults>();
     generator.add::<NestedDefaults>();
-    generator.generate().unwrap();
+    generator
+        .generate(&file, renderers::typescript::TypeScriptRenderer::default())
+        .unwrap();
 
     assert!(file.exists());
     assert_snapshot!(std::fs::read_to_string(file).unwrap());