+++ crates/config/tests/settings_test.rs
@@ -390,12 +390,14 @@ fn generates_typescript() {
     let sandbox = create_empty_sandbox();
     let file = sandbox.path().join("config.ts");
 
-    let mut generator = typescript::TypeScriptGenerator::new(file.clone());
+    let mut generator = schema::SchemaGenerator::default();
     generator.add::<StandardSettings>();
     generator.add::<NestedSettings>();
     generator.add::<NestedVecSettings>();
     generator.add::<NestedMapSettings>();
-    generator.generate().unwrap();
+    generator
+        .generate(&file, renderers::typescript::TypeScriptRenderer::default())
+        .unwrap();
 
     assert!(file.exists());
     assert_snapshot!(std::fs::read_to_string(file).unwrap());