+++ crates/macros/src/config/setting.rs
@@ -109,28 +109,6 @@ impl<'l> Setting<'l> {
         }
     }
 
-    pub fn get_meta(&self, casing_format: &str) -> TokenStream {
-        let optional = self.is_optional();
-
-        let name = self.get_name(Some(casing_format));
-
-        let value = match &self.value_type {
-            SettingType::NestedList { path, .. } => path.to_token_stream().to_string(),
-            SettingType::NestedMap { path, .. } => path.to_token_stream().to_string(),
-            SettingType::NestedValue { path, .. } => path.to_token_stream().to_string(),
-            SettingType::Value { value, .. } => value.to_token_stream().to_string(),
-        }
-        .replace(' ', "");
-
-        quote! {
-            schematic::MetaField {
-                name: #name,
-                kind: #value,
-                optional: #optional,
-            }
-        }
-    }
-
     pub fn get_default_value(&self) -> TokenStream {
         if self.is_optional() {
             quote! { None }