+++ crates/types/src/structs.rs
@@ -4,6 +4,7 @@ use crate::SchemaField;
 pub struct StructType {
     pub additional: bool,
     pub fields: Vec<SchemaField>,
-    pub name: String,
+    pub name: Option<String>,
+    pub partial: bool,
     pub required: Vec<String>,
 }